fixing delete image logic

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -91,14 +91,23 @@ public class ImageController {
 
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
+    //Update paramaters to include HttpSession to pass the session object
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
+        String error = "Only the owner of the image can edit the image";
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        //check if the owners of the picture is same as logged in user as per sesson
+        if ( image.getUser().getUsername() ==  ((User)session.getAttribute("loggeduser")).getUsername() ) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        } else {
+            model.addAttribute("editError", error);
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -139,10 +148,23 @@ public class ImageController {
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
+    //Parameters udpated to pass session object and Model object for verifying user and sending back error code
+
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model, HttpSession session) {
+            String error = "Only the owner of the image can delete the image";
+            Image image = imageService.getImage(imageId);
+            //check if the owners of the picture is same as logged in user as per sesson
+            if ( image.getUser().getUsername() ==  ((User)session.getAttribute("loggeduser")).getUsername() ) {
+                imageService.deleteImage(imageId);
+                return "redirect:/images";
+            } else {
+                model.addAttribute("deleteError", error);
+                model.addAttribute("image", image);
+                model.addAttribute("tags", image.getTags());
+                return "redirect:/images/image";
+            }
+
     }
